Optionally include post content as description
None of my blog posts set the description property, so this field was
previously empty, which meant RSS readers could only display the blog
post’s metadata and link to it. Apparently it’s more customary to
include the whole content in this field (Atom has separate “summary” and
“content” fields), so add a feature to extract the content via Beautiful
Soup and use it as the default description. (Note that simply wrapping
the content in a <div property="description"> wouldn’t work properly –
we want the description to be HTML markup, but this would return its
text content instead.) For now, this is an opt-in feature with the -d /
--description flag, though I could imagine turning it on by default
later.

Also tweak the -k / --keyword description to match the other options
(including the autogenerated -h / --help – otherwise I probably would’ve
preferred making the other descriptions full sentences too instead).